[refactoring] remove unnecessary method and change the store to one
