+++ app/src/main/java/voicebattle/com/shaya/voicebattle/di/StoreModule.kt
@@ -2,22 +2,15 @@ package voicebattle.com.shaya.voicebattle.di
 
 import dagger.Module
 import dagger.Provides
-import voicebattle.com.shaya.voicebattle.meter.AudioStore
 import voicebattle.com.shaya.voicebattle.Dispatcher
-import voicebattle.com.shaya.voicebattle.ranking.FirebaseStore
+import voicebattle.com.shaya.voicebattle.Store
 import javax.inject.Singleton
 
 @Module
 class StoreModule {
     @Provides
     @Singleton
-    fun provideStore(dispatcher: Dispatcher): AudioStore {
-        return AudioStore(dispatcher)
-    }
-
-    @Provides
-    @Singleton
-    fun provideFirebaseStore(dispatcher: Dispatcher): FirebaseStore {
-        return FirebaseStore(dispatcher)
+    fun provideFirebaseStore(dispatcher: Dispatcher): Store {
+        return Store(dispatcher)
     }
 }
