--- app/src/main/java/voicebattle/com/shaya/voicebattle/ranking/FirebaseStore.kt
@@ -1,11 +1,0 @@
-package voicebattle.com.shaya.voicebattle.ranking
-
-import voicebattle.com.shaya.voicebattle.Dispatcher
-import javax.inject.Inject
-
-class FirebaseStore @Inject constructor(val dispatcher: Dispatcher) {
-    var test = dispatcher.on(FirebaseAction.Test::class.java)
-            .map { it.test }
-    var entities = dispatcher.on(FirebaseAction.Ranking::class.java)
-            .map { it.enitities }
-}
