+++ app/src/main/java/voicebattle/com/shaya/voicebattle/ranking/RankingEntity.kt
@@ -2,6 +2,5 @@ package voicebattle.com.shaya.voicebattle.ranking
 
 class RankingEntity(
         var name:String,
-        var power:Long,
-        var ranking:Int) {
+        var power:Long) {
 }
